+++ backend/run.ts
@@ -1,10 +1,10 @@
-import process from "process";
-import open from "open";
+import process from 'process';
+import open from 'open';
 
-import { createFrontend, InjectExpress } from "./app";
-import { Instances, Options } from "./instance";
-import { getLocation, Location, LocationError } from "./location";
-import { getAppInfo, AppInfo, AppInfoError } from "./appInfo";
+import { createFrontend, InjectExpress } from './app';
+import { Instances, Options } from './instance';
+import { getLocation, Location, LocationError } from './location';
+import { getAppInfo, AppInfo, AppInfoError } from './appInfo';
 
 export type Inject = {
   injectFrontend: InjectExpress;
@@ -29,12 +29,12 @@ function actualRun(appInfo: AppInfo, options: Options, inject: Inject): void {
   );
 
   frontend.listen(options.basePort, () => {
-    console.log("Starting webxdc-dev frontend");
+    console.log('Starting webxdc-dev frontend');
   });
 
   instances.start();
 
-  open("http://localhost:" + options.basePort);
+  open('http://localhost:' + options.basePort);
 }
 
 export function run(locationStr: string, options: Options, inject: Inject) {
@@ -49,13 +49,13 @@ export function run(locationStr: string, options: Options, inject: Inject) {
     throw e;
   }
 
-  for (const signal in ["SIGINT", "SIGTERM"]) {
+  for (const signal in ['SIGINT', 'SIGTERM']) {
     process.on(signal, () => {
       location.dispose();
     });
   }
 
-  console.log("Starting webxdc project in:", locationStr);
+  console.log('Starting webxdc project in:', locationStr);
 
   getAppInfo(location)
     .then((appInfo) => {